+++ launcher-core/src/main/java/com/launcher/core/launch/LaunchContext.java
@@ -1,6 +1,7 @@
 package com.launcher.core.launch;
 
 import com.launcher.core.configuration.LauncherConfiguration;
+import com.launcher.core.download.DownloadPlan;
 import com.launcher.core.verification.model.VerificationPlan;
 import com.launcher.model.manifest.Manifest;
 
@@ -9,6 +10,7 @@ public class LaunchContext {
     private Manifest manifest;
     private final LauncherConfiguration launcherConfiguration;
     private VerificationPlan verificationPlan;
+    private DownloadPlan downloadPlan;
 
     public LaunchContext(LauncherConfiguration launcherConfiguration) {
         this.launcherConfiguration = launcherConfiguration;
@@ -29,6 +31,15 @@ public class LaunchContext {
 
     public LaunchContext setVerificationPlan(VerificationPlan verificationPlan) {
         this.verificationPlan = verificationPlan;
+        return this;
+    }
+
+    public DownloadPlan getDownloadPlan() {
+        return downloadPlan;
+    }
+
+    public LaunchContext setDownloadPlan(DownloadPlan downloadPlan) {
+        this.downloadPlan = downloadPlan;
         return this;
     }